Subclass torch
