+++ opendg/loader/__init__.py
@@ -1,3 +1,2 @@
-from opendg.loader.base import DGBaseLoader
 from opendg.loader.dataloader import DGDataLoader
 from opendg.loader.neighbor_loader import DGNeighborLoader
