+++ opendg/loader/neighbor_loader.py
@@ -1,45 +1,36 @@
 from __future__ import annotations
 
-from typing import List
+from typing import Any, List
 
 from opendg._storage import DGSliceTracker
 from opendg.graph import DGraph
-from opendg.loader.base import DGBaseLoader
+from opendg.loader.dataloader import DGDataLoader
 
 
-class DGNeighborLoader(DGBaseLoader):
+class DGNeighborLoader(DGDataLoader):
     r"""Load data from DGraph using a memory based sampling function.
 
     Args:
         dg (DGraph): The dynamic graph to iterate.
         num_nbrs (List[int]): Number of neighbors to sample at each hop (-1 to keep all)
-        batch_size (int): The batch size to yield at each iteration.
-        batch_unit (str): The unit corresponding to the batch_size.
-        drop_last (bool): Set to True to drop the last incomplete batch.
-        materialize (bool): If True (default), auto-materialize the DGraph before yielding a batch.
+        **kwargs (Any): Additional arguments to the DGDataLoader
 
     Raises:
         ValueError: If the num_nbrs list is empty.
-        ValueError: If the batch_unit and dg time unit are not both ordered or both not ordered.
-        ValueError: If the batch_unit and dg time unit are both ordered but the graph is coarser than the batch.
     """
 
     def __init__(
         self,
         dg: DGraph,
         num_nbrs: List[int],
-        batch_size: int = 1,
-        batch_unit: str = 'r',
-        drop_last: bool = False,
-        materialize: bool = True,
+        **kwargs: Any,
     ) -> None:
         if not len(num_nbrs):
             raise ValueError('num_nbrs must be non-empty')
         if not all([isinstance(x, int) and (x == -1 or x > 0) for x in num_nbrs]):
             raise ValueError('Each value in num_nbrs must be a positive integer or -1')
         self._num_nbrs = num_nbrs
-        self._ordered = batch_unit == 'r'
-        super().__init__(dg, batch_size, batch_unit, drop_last, materialize)
+        super().__init__(dg, **kwargs)
 
     @property
     def num_hops(self) -> int:
@@ -61,9 +52,9 @@ class DGNeighborLoader(DGBaseLoader):
             for node, _ in seed_nbrhood[-1]:  # Only care about final hop
                 temporal_nbrhood.add(node)  # Don't care about time info either
 
-        if self._ordered:
-            batch = self._dg.slice_events(end_idx=batch._slice.start_idx)
+        if self._iterate_by_time:
+            batch = self._dg.slice_time(end_time=batch.end_time)
         else:
-            batch = self._dg.slice_time(end_time=batch.start_time)
+            batch = self._dg.slice_events(end_idx=batch._slice.end_idx)
         batch = batch.slice_nodes(list(temporal_nbrhood))
         return batch
